+++ src/sys/env.rs
@@ -110,6 +110,8 @@ cfg_sel!({
         target_arch = "arm",
         target_arch = "riscv32",
         target_arch = "riscv64",
+        target_arch = "loongarch32",
+        target_arch = "loongarch64",
         all(target_arch = "xtensa", feature = "openocd-semihosting"),
     ))]
     {